telegram_boot/requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.0
paho-mqtt==2.1.0
requests==2.31.0